major refactors
