+++ internal/vacancy/handler.go
@@ -1,36 +1,32 @@
-package handler
+package vacancy
 
 import (
 	"html/template"
 	"job-finder/internal/client/rapid"
+	"job-finder/internal/middleware"
+	"job-finder/internal/models"
 	"log"
 	"net/http"
 	"path/filepath"
-
-	"github.com/joho/godotenv"
 )
 
 type TemplateData struct {
-	Registration string
-	Query        string
-	Jobs         []rapid.Job
+	User  *models.User
+	Query string
+	Jobs  []rapid.Job
 }
 
 func JobHandler(w http.ResponseWriter, r *http.Request) {
-	err := godotenv.Load()
-	if err != nil {
-		log.Fatal("Error loading .env file")
-	}
+	user, _ := middleware.GetUserFromContext(r.Context())
 
 	query := r.URL.Query().Get("q")
-	register := r.URL.Query().Get("register")
 
 	jobs := rapid.GetJob(query, "", "", "", "", "", "", "", "", "")
 
 	data := TemplateData{
-		Registration: register,
-		Query:        query,
-		Jobs:         jobs,
+		User:  user,
+		Query: query,
+		Jobs:  jobs,
 	}
 
 	tmpl, err := template.ParseFiles(filepath.Join("web", "templates", "index.html"))
